added getPost() function to use on all the needed pages
