+++ pages/index.js
@@ -1,10 +1,7 @@
-import fs from 'fs';
-import path from 'path';
-import matter from 'gray-matter';
 import Layout from '@/components/Layout';
 import Link from 'next/link';
 import Post from '@/components/Post';
-import {sortByDate} from '@/utils/index';
+import { getPosts } from '@/lib/posts';
 
 export default function HomePage({posts}) {
   return (
@@ -26,31 +23,12 @@ export default function HomePage({posts}) {
 
 export async function getStaticProps(){
 
-  const files = fs.readdirSync(path.join('data/posts'))
-  //console.log(files) //this to show the files in the dir
 
-
-  // map thru the files array and create a new array with the slug and the frontmatter data
-  const posts = files.map(filename => {
-      const slug = filename.replace('.md', '');
-
-      const markdownWithMeta = fs.readFileSync(path.join('data/posts', filename), 'utf-8');
-
-      //Using gray-matter package to turn the string of the frontmatter to object
-      //We are renaming the data object to frontmatter
-      const {data:frontmatter} = matter(markdownWithMeta)
-
-      return {
-          slug,
-          frontmatter,
-      }
-  })
+  
 
   return {
       props: {
-        //return the post but after you sort by date
-        // and only return the latest 6 posts
-          posts: posts.sort(sortByDate).slice(0,6),
+          posts: getPosts().slice(0,6),
       },
   }
 }
